Projects/X-CUBE-AI/models/GenHeader/requirements.txt
hydra-core
omegaconf
numpy>=1.20
munch
seaborn
librosa
